30 - Presentation Layer : Introduction

+++ lib/movies/data/models/movies_model.dart
@@ -17,9 +17,7 @@ class MoviesModel extends Movies {
         backdropPath: json["backdrop_path"],
         genreIDs: List<int>.from(json["genre_ids"].map((e) => e)),
         overview: json["overview"],
-
-        /// TODO : CHECK THIS
-        voteAverage: json["vote_average"],
+        voteAverage: json["vote_average"].toDouble(),
         releaseDate: json["release_date"],
       );
 }
